free: silence a false positive in check_free
The problem here is you have code like:

void my_free(void *p)
{
	if (p)
		kfree(p);
}

The issue is, do we try to break up the return states and say that this
will free it on the path where p is non-NULL?  Sure, that sounds nice but
it's tricky to do.

So what Smatch does instead is it says that if we free the pointer on
the non-NULL path, and leave it alone on the NULL path then the pointer
is not &merged but actually just &freed.

Which seems reasonable.  But it becomes less reasonable the further you
move away from the NULL test.  So what happened is, I guess the database
got out of sync, and it said that "We always pass a NULL pointer to this
function. And one path frees the pointer, so that means that all paths
free the pointer."  Which was wrong.

The fix here is that if we're merging function returns and some free it
and some don't then don't treat it as a free function.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>