extra: create more fake assignments in returned states
Smatch creates a bunch of fake assignments.  If you return a parameter
then that's faked.  If you assign a parameter to another parameter.  But
what wasn't handled is if you do "ret->member = param;"

This adds support for that.  Now that kind of thing will generate a
fake assignment.

It's a bit weird that this is handled in smatch_extra.c instead of
smatch_function_hooks.c but I did that because that's where PARAM_VALUE
is handled.

Also it's a bit of a mess that some of these end up as "[== $0]" and some
end up as "[$0->foo]".  It should be made more consistent.  #TODO

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>