param_limit: use the orig_name and sym
This code does something weird and possibly stupid.  It goes through
all the extra states and then checks if:

1) it's a parameter
2) the state has been limited since the start of the function

But the problem is we could have done a "x = param; if (x < 0) return;"
thing which limits x and x is sort of a parameter.  But instead of looking
up the original value of "param" it looks up the original value of "x"
which was none.  Look up the original value of "param" instead.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>