refcount_info: add support for find_process()
There was a refcount leak related to find_process().  Smatch doesn't
parse it correctly because it takes the reference if you pass "true" to
the "ref" param, and doesn't take it if you don't.

This is not a common function, but I just wrote this code as a
demonstration of how to handle that.

(I have the idea of creating mini-strees which would handle this kind of
thing automatically but that's quite a complicated process).

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>