extra: fix a bug incrementing unsigned zeros
Imagine you have "x++".  First of all Smatch doesn't handle loops
correctly, so if we're in a loop we just mark it as unknown.  Then
if x is "s32min-3" we want the result to be "s32min-4".  So that's
why we test against sval_is_min().  But there is the case where
x is zero and that's a minimum but we want the result to be 1.

Ugh...  This is just a mess of hacks.  :/

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>