conditions: handle conditions like "if ((x - y) != 0) {"
Sometimes when we're doing a circular loop then the
"nr_present - nr_read" is the remaining items or whatever.  So it's
kind of natural to write code like if ((nr_present - nr_read) != 0) {.
But that's the equivalent of writing if (nr_present != nr_read) and
it's easier if we always express that comparison in the same way
instead of making the checks handle it as a special case.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>