comparison: handle "a = b - 1"
If you have "a = b - 1" then you would think that "a < b" but the code
to implement this in Smatch would only say that was true if we were
sure that it didn't underflow.  This led to a number of false positives
because there is a lot of code which does "a - 1", knowing full well that
it can't underflow.  Just silence these false positives by assuming that
there isn't an underflow.  In other words, for this specific case, err on
the side of ignoring false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>