flow: handle foo[i++] = a ? b : c;
If we have an assignment "foo[i++] = a ? b : c;" then what Smatch does
is it creates something like this:

if (a)
    foo[i++] = b;
else
    foo[i++] = c;

The problem is that we parse "i++;" multiple times and then in the
caller we parse the left side of the assignment again.

We add the i++ to the post op stack twice, when we're parsing
foo[i++] = b; and foo[i++] = c; and then the foo[i++].  So to
fix this, we first add it and then when we finish parsing
"foo[i++] = b;" we process the post op stack.  Then we do the same
for the "foo[i++] = c;" and then in the caller, we say don't parse the
left hand side.

The changes to parse_assignment() are messier than I had wanted.  A bit
of a hack.  Possibly each of the handle__builtin_choose_expr_assigns()
functions should parse the left hand side.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>